requirements.txt
# Core dependencies
fastapi
uvicorn[standard]
pydantic>=2.0
python-multipart

# LangChain and LangGraph
langchain
langchain-community
langchain-core
langgraph
langgraph-checkpoint-sqlite

# LLM providers (choose based on your needs)
openai
anthropic

# Document processing
python-docx
PyPDF2

# Utilities
python-dotenv
pathlib